user has access to app if they belong to any app roles
