+++ Lib/XTI_App.Api/IAppApiUser.cs
@@ -4,6 +4,7 @@ namespace XTI_App.Api
 {
     public interface IAppApiUser
     {
+        Task<bool> HasAccessToApp();
         Task<bool> HasAccess(ResourceAccess resourceAccess, AccessModifier modifier);
     }
 }
